backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.0
python-multipart==0.0.6
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
moviepy==1.0.3
imageio-ffmpeg==0.4.9
numpy==1.26.0
Pillow==10.1.0
decorator==4.4.2 